style(code): Convert main code to ES6 compatible with Node 4
