+++ .eslintrc
@@ -10,7 +10,16 @@
       "printWidth": 80
     }],
     "no-underscore-dangle": 0,
-    "no-plusplus": 0
+    "no-plusplus": 0,
+    "no-use-before-define": 0,
+    "func-names": 0,
+    "no-shadow": 0,
+    "consistent-return": 0,
+    "no-restricted-syntax": 0,
+    "no-multi-assign": 0,
+    "prefer-rest-params": 0,
+    "no-param-reassign": 0,
+    "strict": 0
   },
   "env": {
     "mocha": true
