+++ lib/bulker.js
@@ -1,4 +1,6 @@
-var EventEmitter = require('events').EventEmitter;
+'use strict';
+
+const EventEmitter = require('events').EventEmitter;
 
 module.exports = Bulker;
 
@@ -13,18 +15,18 @@ module.exports = Bulker;
  * @event `sent`: Emitted when bulk has sent a buffer.
  */
 function Bulker(client, options) {
-  options = options || {};
+  options = options || {}; // eslint-disable-line
 
-  var timeout;
-  var self = this;
-  var delay = options.delay || 1000;
-  var size = options.size || 1000;
-  var buffer = [];
+  let timeout;
+  const self = this;
+  const delay = options.delay || 1000;
+  const size = options.size || 1000;
+  let buffer = [];
 
   self.push = function() {
-    var sending = false;
+    let sending = false;
     if (arguments.length) {
-      buffer.push.apply(buffer, arguments);
+      buffer.push(...arguments);
       sending = buffer.length >= size;
       if (sending) {
         self.flush();
@@ -38,7 +40,7 @@ function Bulker(client, options) {
   self.delay = function() {
     clearTimeout(timeout);
     timeout = setTimeout(
-      function() {
+      () => {
         self.flush();
       },
       delay
@@ -50,10 +52,10 @@ function Bulker(client, options) {
   };
 
   self.flush = function() {
-    var len = buffer.length;
+    const len = buffer.length;
     clearTimeout(timeout);
     if (len) {
-      client.bulk({ body: buffer }, function(err) {
+      client.bulk({ body: buffer }, err => {
         if (err) {
           self.emit('error', err);
         } else {
