+++ lib/client.js
@@ -1,7 +1,9 @@
-var elasticsearch = require('elasticsearch');
+'use strict';
+
+const elasticsearch = require('elasticsearch');
 
 module.exports = function(options) {
-  var opts = {};
+  const opts = {};
 
   if (Array.isArray(options.hosts)) {
     opts.host = options.hosts;
