+++ lib/mapping.js
@@ -1,5 +1,4 @@
-// Old version of node does not support Object.assign
-var extend = Object.assign || require('util')._extend;
+'use strict';
 
 module.exports = {};
 
@@ -12,22 +11,26 @@ module.exports.generate = generate;
  * @returns {Object}
  */
 function generate(schema, version) {
-  var mapping = {};
-  var explicit = hasExplicit(schema);
-  var defaultTypes = getDefault(version);
+  const mapping = {};
+  const explicit = hasExplicit(schema);
+  const defaultTypes = getDefault(version);
 
-  Object.keys(schema.paths).forEach(function(name) {
+  Object.keys(schema.paths).forEach(name => {
     // ignore _id because is used as index
     if (name === '_id') {
       return;
     }
 
-    var path = schema.paths[name];
-    var type = (path.caster && path.caster.instance
+    const path = schema.paths[name];
+    const type = (path.caster && path.caster.instance
       ? path.caster.instance
       : path.instance).toLowerCase();
 
-    if (explicit && isEmbedded(type) && !path.options.hasOwnProperty('es_indexed')) {
+    if (
+      explicit &&
+      isEmbedded(type) &&
+      !({}).hasOwnProperty.call(path.options, 'es_indexed')
+    ) {
       path.options.es_indexed = hasExplicit(path.schema);
     }
 
@@ -35,15 +38,15 @@ function generate(schema, version) {
       return;
     }
 
-    var current = mapping;
-    var names = name.split('.');
+    let current = mapping;
+    const names = name.split('.');
 
     // handle plain object
     if (names.length > 1) {
-      names.forEach(function(name, index) {
+      names.forEach((name, index) => {
         if (index === names.length - 1) {
           // last item is the target
-          current = current[name] = { type: type };
+          current = current[name] = { type };
         } else {
           if (!current[name]) {
             current[name] = { type: 'object', properties: {} };
@@ -52,14 +55,17 @@ function generate(schema, version) {
         }
       });
     } else {
-      current = mapping[name] = { type: type };
+      current = mapping[name] = { type };
     }
 
     if (path.options.es_type && typeof path.options.es_type === 'object') {
       current.type = 'object';
       current.properties = generateESTypeMapping(path.options.es_type);
     } else {
-      if (!path.options.hasOwnProperty('es_value') || !path.options.es_type) {
+      if (
+        !({}).hasOwnProperty.call(path.options, 'es_value') ||
+        !path.options.es_type
+      ) {
         if (isEmbedded(type)) {
           current.type = 'object';
           current.properties = generate(path.schema, version);
@@ -70,14 +76,14 @@ function generate(schema, version) {
       }
 
       // propagate es_ options from schema to mapping
-      Object.keys(path.options).forEach(function(key) {
+      Object.keys(path.options).forEach(key => {
         if (key !== 'es_indexed' && key.substr(0, 3) === 'es_') {
           current[key.substr(3)] = path.options[key];
         }
       });
     }
 
-    if (path.options.hasOwnProperty('es_value')) {
+    if (({}).hasOwnProperty.call(path.options, 'es_value')) {
       current.value = typeof path.options.es_value === 'function'
         ? path.options.es_value
         : (function() {
@@ -89,23 +95,26 @@ function generate(schema, version) {
   delete mapping[schema.get('versionKey')];
 
   if (schema.options && schema.options.es_extend) {
-    extend(mapping, generateESTypeMapping(schema.options.es_extend, true));
+    Object.assign(
+      mapping,
+      generateESTypeMapping(schema.options.es_extend, true)
+    );
   }
 
   return mapping;
 }
 
 function isEmbedded(type) {
-  return type === 'embedded' || type === 'array' || type === 'mixed';
+  return type === 'embedded' || type === 'array'; // || type === 'mixed';
 }
 
 function hasExplicit(schema) {
-  return schema && schema.paths ? Object.keys(schema.paths).some(function(name) {
+  return schema && schema.paths ? Object.keys(schema.paths).some(name => {
         if (name === '_id') {
-          return;
+          return; // eslint-disable-line
         }
-        var path = schema.paths[name];
-        var type = (path.caster && path.caster.instance
+        const path = schema.paths[name];
+        const type = (path.caster && path.caster.instance
           ? path.caster.instance
           : path.instance).toLowerCase();
         if (isEmbedded(type)) {
@@ -113,31 +122,36 @@ function hasExplicit(schema) {
             return true;
           }
         }
-        return path.options.hasOwnProperty('es_indexed');
+        return ({}).hasOwnProperty.call(path.options, 'es_indexed');
       }) : false;
 }
 
 function generateESTypeMapping(content, esExtendMode) {
-  var properties = {};
+  const properties = {};
   // browse properties
-  Object.keys(content).forEach(function(key) {
+  Object.keys(content).forEach(key => {
     if (content[key] && typeof content[key] === 'object') {
       // only browse well formed object
       properties[key] = {};
       if (content[key].es_type && typeof content[key].es_type === 'object') {
         properties[key].type = 'object';
-        properties[key].properties = generateESTypeMapping(content[key].es_type, esExtendMode);
+        properties[key].properties = generateESTypeMapping(
+          content[key].es_type,
+          esExtendMode
+        );
       } else {
-        Object.keys(content[key]).forEach(function(subkey) {
-          var targetSubkey = subkey.replace(/^es\_/, ''); // remove plugin prefix
-          var value = content[key][subkey];
-          var original = value;
+        Object.keys(content[key]).forEach(subkey => {
+          const targetSubkey = subkey.replace(/^es_/, ''); // remove plugin prefix
+          let value = content[key][subkey];
+          const original = value;
 
           // build a function to be ignored in the mapping sent
           if (subkey === 'es_value' && esExtendMode) {
             value = function(_, context) {
               // serialised function of es_value handle prototype (value, context) where value is always undefined in options.es_extend feature
-              return typeof original === 'function' ? original(context.document) : original;
+              return typeof original === 'function'
+                ? original(context.document)
+                : original;
             };
           }
