+++ lib/utils.js
@@ -1,5 +1,8 @@
-var mongoose = require('mongoose');
-var _Promise = mongoose.Promise.ES6;
+'use strict';
+
+const mongoose = require('mongoose');
+
+const _Promise = mongoose.Promise.ES6;
 
 module.exports = {};
 
@@ -13,10 +16,10 @@ module.exports.run = function(callback, fn) {
   if (typeof callback === 'function') {
     try {
       fn(
-        function(result) {
+        result => {
           callback(undefined, result);
         },
-        function(reason) {
+        reason => {
           callback(reason);
         }
       );
@@ -43,9 +46,9 @@ module.exports.lcFirst = function(str) {
  * @returns {Object}
  */
 module.exports.highClone = function(source) {
-  var result = {};
+  const result = {};
   if (source) {
-    Object.keys(source).forEach(function(key) {
+    Object.keys(source).forEach(key => {
       result[key] = source[key];
     });
   }
@@ -59,9 +62,10 @@ module.exports.highClone = function(source) {
  * @returns {Array}
  */
 module.exports.getUndefineds = function getUndefineds(document, mapping) {
-  var field, result = [];
+  let field;
+  const result = [];
   for (field in mapping.properties) {
-    if (mapping.properties.hasOwnProperty(field)) {
+    if (({}).hasOwnProperty.call(mapping.properties, field)) {
       if (document[field] === undefined && document.isModified(field)) {
         result.push(field);
       }
@@ -78,19 +82,23 @@ module.exports.getUndefineds = function getUndefineds(document, mapping) {
  * @returns {Object}
  */
 module.exports.serialize = function serialize(document, mapping, main) {
-  var name;
+  let name;
 
-  main = main || document;
+  main = main || document; // eslint-disable-line
 
   function _serializeObject(object, mappingData) {
-    var serialized = {};
-    Object.keys(mappingData.properties).forEach(function(field) {
-      var value;
-      var property = mappingData.properties[field];
+    const serialized = {};
+    Object.keys(mappingData.properties).forEach(field => {
+      let value;
+      const property = mappingData.properties[field];
       try {
-        if (property.hasOwnProperty('value')) {
+        if (({}).hasOwnProperty.call(property, 'value')) {
           value = typeof property.value === 'function'
-            ? property.value(object[field], { document: main, container: object, field: field })
+            ? property.value(object[field], {
+                document: main,
+                container: object,
+                field,
+              })
             : property.value;
         } else if (object[field] !== undefined) {
           value = serialize.call(object, object[field], property, main);
@@ -113,9 +121,7 @@ module.exports.serialize = function serialize(document, mapping, main) {
 
   if (mapping.properties && document) {
     if (Array.isArray(document)) {
-      return document.map(function(object) {
-        return _serializeObject(object, mapping);
-      });
+      return document.map(object => _serializeObject(object, mapping));
     }
     return _serializeObject(document, mapping);
   }
